apps/api/src/agents/python/requirements.txt
browser-use
langchain-google-genai
langchain-anthropic
python-dotenv
pydantic>=2.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
langchain-openai==0.2.0
httpx==0.27.2
